Remunge the package installation situation.
So, I'm currently experiencing the problem of the openldap-servers
package deleting /etc/openldap/slapd.conf whenever it's installed. I'm
not a gigantic fan of things operating like this, but it's correctly
fixed by correct ordering of installation.

So: split the $openldap_packages list to only be for servers, and not
have common packages in the list. That way we don't have to do silly
things with virtual packages and realization to ensure there aren't
conflicts with the client packages. Then the server package dependancy
order actually works.

Incidentally, this also leads to ruby-ldap dumping core when it finishes
making changes to ldap. This is less than ideal. However I've determined
it doesn't actually affect puppets operation as it only happens when
everything gets unloaded on exit().

Clearly this isn't something we should accept in the long term, but I
can't fix it right now.